Update "system-appearance" patch and related docs
The implementation of this feature now relies on Key-Value Observing to
detect system appearance changes, rather than the
`viewDidChangeEffectiveApperance:` NSView method.

The old implementation ran the `ns-system-appearance-change-functions`
hook every time a new NSView was created, so in Emacs terms, every time
a new frame or child frame was created. KVO on the NSApp (EmacsApp)
instance allows handling this change at the application level rather
than the view level.
The hook is also always run now, even if it is empty; This has the
benefit of removing the requirement of adding functions to the hook in
`early-init.el` (instead of `init.el`) in order for them to be run when
Emacs starts.

Fixes #240.