fix(native-comp): Add libgccjit paths to build environment
When installed with Homebrew, libgccjit 12 writes header files to the Homebrew
`include` directory and DLLs to Homebrew's `lib/gcc/current` directory. If these
locations are not added to CPATH and LIBRARY_PATH, respectively, the Emacs build
toolchain is unable to locate libgccjit, leading to the build failure described
in #485.

This change appends those libgccjit paths to CPATH and LIBRARY_PATH for all
Emacs versions that support the native compilation build flag.

Resolves: #485