fix(build): Add grep as build-time dependency
I found that on my system, I had a different incompatible version of grep from a
plan9 installation on my PATH that was superseding the builtin grep on one of my
Macs. This explains why the `with-native-comp` build was failing on one of my
Macs but not the other - the Mac with the successful emacs-plus@28
`with-native-comp` build only had the built-in version of grep on its PATH.

Adding GNU grep as a build-time dependency will include it in the `gnubin`
directory, which is prepended to the PATH in each formula. I've also added an
explicit PATH prepending step for each of the GNU build-time dependencies. This
indicates the coupling of those dependencies with the `gnubin` PATH prepending.

Related to d12frosted/homebrew-emacs-plus#555 and d12frosted/homebrew-emacs-plus#556