Turn native comp on by default in emacs 30+ (#784)
* always build with aot native comp for emacs 30+

Emacs 30 defaults to 'yes' for the native compilation flag.  Previous
version defaulted to 'no'.  This formula has been opinionated about
forcing the 'aot' value for users who opt-in to native compilation.
This change aligns with Emacs 30 by turning native compilation on by
default, and carries the opinion forward of setting the flag's value to
'aot'.

Ref: #717
Close: #782

* docs: remove outdated reference to no-frame-refocus

Ref: #611